refactor(vcd): rename tracking file extension
ref: #MANAGER-15832
