+++ packages/manager/apps/hpc-vmware-managed-vcd/src/components/tiles/organization-data-tile/backup-tile-item/BackupTileItem.component.tsx
@@ -18,7 +18,7 @@ import {
   getBackupBadgeStatus,
 } from '@/utils/veeamBackupBadge';
 import TEST_IDS from '@/utils/testIds.constants';
-import { TRACKING } from '@/tracking.constant';
+import { TRACKING } from '@/tracking.constants';
 
 type TTileProps = {
   vcdOrganization: VCDOrganization;
