+++ packages/manager/apps/hpc-vmware-managed-vcd/src/components/tiles/organization-general-information-tile/OrganizationGeneralInformationTile.spec.tsx
@@ -11,7 +11,7 @@ import {
 import userEvent from '@testing-library/user-event';
 import OrganizationGeneralInformationTile from './OrganizationGeneralInformationTile.component';
 import TEST_IDS from '../../../utils/testIds.constants';
-import { TRACKING } from '../../../tracking.constant';
+import { TRACKING } from '../../../tracking.constants';
 
 const trackClickMock = vi.fn();
 vi.mock('@ovh-ux/manager-react-shell-client', async (importOriginal) => {
