+++ packages/manager/apps/hpc-vmware-managed-vcd/src/components/tiles/organization-options-tile/OrganizationOptionsTile.component.tsx
@@ -7,7 +7,7 @@ import { OdsText } from '@ovhcloud/ods-components/react';
 import { ODS_BUTTON_VARIANT, ODS_ICON_NAME } from '@ovhcloud/ods-components';
 import { useOvhTracking } from '@ovh-ux/manager-react-shell-client';
 import { WINDOWS_LICENSE_PLANCODE } from '@/utils/planCode.constants';
-import { TRACKING } from '@/tracking.constant';
+import { TRACKING } from '@/tracking.constants';
 import TEST_IDS from '@/utils/testIds.constants';
 
 export default function OrganizationOptionsTile({
