+++ packages/manager/apps/hpc-vmware-managed-vcd/src/pages/onboarding/Onboarding.page.tsx
@@ -5,7 +5,7 @@ import { Card, OnboardingLayout } from '@ovh-ux/manager-react-components';
 import { useOvhTracking } from '@ovh-ux/manager-react-shell-client';
 import useGuideUtils from '@/hooks/guide/useGuideUtils';
 import vmwareBroadcomOVHCloud from '@/assets/VmwareBroadcomxOVHcloud.svg';
-import { TRACKING } from '@/tracking.constant';
+import { TRACKING } from '@/tracking.constants';
 
 export default function Onboarding() {
   const { t } = useTranslation('onboarding');
